+++ org.mixedrealitytoolkit.theming/ItemData/BaseThemeItemData.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Mixed Reality Toolkit Contributors
+// Licensed under the BSD 3-Clause
+
+using UnityEngine;
+
+namespace MixedReality.Toolkit.Theming
+{
+    public abstract class BaseThemeItemData<T>
+    {
+        [field: SerializeField]
+        public T Value { get; set; }
+    }
+}